CHANGELOG: commit to update dependencies and set naming convention for namespace 1. refactor drivers directory with standarized naming convention:    - drivers should be in directories of corresponding companies    - subdirectories should be company names    - subdirectories should be all lowercase    - underscores should be used for spaces    - hyphens only allowed when they are part of the company name (e.g. mini-circuits)    - separate acronyms from whole worlds in company names (e.g. aa_opto_electronic, bk_precision) 2. complete rewrite of setup.py using new lantz-core setup.py and add pyserial and pyusb dependencies, update general settings
