+++ lantz/drivers/allied_vision/tests/vimbatest.py
@@ -1,4 +1,4 @@
-from lantz.drivers.alliedvision import list_cameras
+from lantz.drivers.allied_vision import list_cameras
 
 if __name__ == '__main__':
     print(list_cameras())
